CCD-1267 Continuing additions of tests for test_bestrefs.py
