+++ CHANGES.rst
@@ -4,9 +4,15 @@
 General
 -------
 
+- Created pytest for bestrefs.py
+
+11.17.01 (unreleased)
+=====================
+
+General
+-------
+
 - Removed python 3.8 check from ci.yml [#934]
-
-
 
 
 11.17.0 (2023-04-21)
