--- pytest.ini
@@ -1,4 +1,0 @@
-[pytest]
-python_files = test_*
-python_classes = *Tests
-python_functions = test_*
